:feat Major Update
New way on creating networks, also differents components configuration and added a menu for play with activated Trial Networks
