+++ package.json
@@ -9,16 +9,14 @@
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@mui/material": "^6.4.8",
-    "@radix-ui/react-dialog": "^1.1.6",
+    "@tailwindcss/typography": "^0.5.16",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
     "axios": "^1.7.9",
     "d3": "^7.9.0",
-    "dotenv": "^16.4.7",
-    "framer-motion": "^12.5.0",
     "frontend": "file:",
     "fs": "^0.0.1-security",
+    "highlight.js": "^11.11.1",
     "js-yaml": "^4.1.0",
     "jwt-decode": "^4.0.0",
     "nth-check": "2.1.1",
@@ -29,9 +27,8 @@
     "react-router-dom": "^7.1.1",
     "react-scripts": "5.0.1",
     "react-syntax-highlighter": "^15.6.1",
-    "rehype-raw": "^7.0.0",
+    "rehype-highlight": "^7.0.2",
     "remark-gfm": "^4.0.1",
-    "url": "^0.11.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
