--- src/auxFunc/examples.txt
@@ -1,23 +1,0 @@
----------------SOME EXAMPLES OF SPECIAL CASES--------------------------
-
-LoadcoreAgent => Checkbox example
-OneKE => Double checkbox example
-NokiaRadio => Select example
-NokiaRadio => Example of validation and field generation
-Stf_UE => Validation example without special fields
-Ueransim=> Example of double select and checkbox
-Ueransim => Example of special required fields
-Ueransim => Special fields includes  special handling
-Berlin_ran => Double select
-Iswireless_radio => Simple select
-Ixc_endpoint => Simple checkbox
-Open5gs_VM => Doubles chechbox
-Open5gs_k8s => Simple select
-
--------------EXAMPLES OF ERROR VALDIATION--------------------------
-Ks8500Runner, OCF => Valid URL example
-OneKE => Valid IP list, IP range and boolean value
-Open5GS => Valid IP, number length, addition check
-STF_UE (almost everyone) => Integer Validation
-Tn_Init, Tn_Vxlan, Vnet (Equals???) => Validation of IP and IP lists
-Ixc_endpoint => Valid IP
