--- src/components/library/Int_p4_sw.jsx
@@ -1,1 +1,0 @@
-//TODO LATER BECOUSE THE API DOESNT GIVE THE CORRECT FIELDS
