--- src/components/library/Xrext.jsx
@@ -1,289 +1,0 @@
-import { faTrash } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import React, { useEffect, useRef, useState } from "react";
-
-const Xrext = ({ id, removeComponent, onChange, list, defaultValues, name, request }) => {
-  const [data, setData] = useState(null);
-  const [formValues, setFormValues] = useState({});
-  const [errorMessages, setErrorMessages] = useState({});
-  const [requiredFields, setRequiredFields] = useState({});
-  const hasFetched = useRef(false);
-
-  useEffect(() => {
-    const loadData = async () => {
-      if (hasFetched.current) return;
-      hasFetched.current = true;
-      let result= null;
-      result = await request;
-      if (!result) {
-        hasFetched.current = false;
-        return;
-      }
-      if (result) {
-        setData(result.component_input);
-        const deps=[];
-        const required = [];  // Array to store the required fields
-        // Initialize form values with default values
-        const initialValues = {};
-        for (const key in result.component_input) {
-          const field = result.component_input[key];
-          
-          // No default values if the field is special type
-          if (field.type === "str" || field.type === "int" || field.type === "bool") {
-            initialValues[key] = field.default_value || "";
-            if (defaultValues && key in defaultValues){
-              initialValues[key] = defaultValues[key];
-            }
-          } else {
-            initialValues[key] ="";
-            deps.push(key);
-            if (defaultValues && key in defaultValues){
-              initialValues[key] = defaultValues[key];
-            }
-          }
-          
-          if (field.required_when === true) {
-            required.push(key);
-          }
-        }
-        required.push("name");
-        name ? initialValues['name'] = name : initialValues['name'] = '';
-        initialValues['required']=required;
-        initialValues['dependencies']=deps;
-        setFormValues(initialValues);
-        setRequiredFields(required);
-        // Call onChange for each initial value
-        for (const key in initialValues) {
-          onChange(id, key, initialValues[key]);
-        }
-      }
-    };
-    loadData();
-  }, [id, onChange, defaultValues, name, request]);
-
-  useEffect(() => {
-        // Make sure that "one_xrext_networks" is an array
-        const networks = Array.isArray(formValues["one_xrext_networks"])
-          ? formValues["one_xrext_networks"]
-          : [];
-      
-        // Filter the networks that are included in the list
-        const validNetworks = networks.filter((network) => list.includes(network));
-      
-        // If the length of the filtered networks is different from the original length
-        if (validNetworks.length !== networks.length) {
-          setFormValues((prevState) => ({
-            ...prevState,
-            "one_xrext_networks": validNetworks,
-          }));
-      
-          // Call onChange to update the state in the parent component
-          onChange(id, "one_xrext_networks", validNetworks);
-        }
-      }, [list,formValues,id,onChange]);
-
-    const handleCheckboxChange = (event, key, network) => {
-      const updatedNetworks = event.target.checked
-        ? [...formValues[key], network]
-        : formValues[key].filter((n) => n !== network);
-    
-      // Update the form values with the new list of networks
-      setFormValues((prevState) => ({
-        ...prevState,
-        [key]: updatedNetworks,
-      }));
-    
-      // call onChange to update the state in the parent component
-      onChange(id, key, updatedNetworks);
-    };
-
-  const handleChange = (event) => {
-    const { name, value } = event.target;
-
-    // Update the form values with the user input
-    setFormValues((prevState) => ({
-      ...prevState,
-      [name]: value,
-    }));
-
-    // Call the onChange function to update the parent state
-    onChange(id, name, value);
-
-    // Field validation
-    if (requiredFields.includes(name)) {
-      if (value.trim() === "") {
-        setErrorMessages((prevState) => ({
-          ...prevState,
-          [name]: `${name} cannot be empty.`,
-        }));
-      } else {
-        setErrorMessages((prevState) => {
-          const newState = { ...prevState };
-          delete newState[name]; // Delete the error message if the field is not empty
-          return newState;
-        });
-      }
-    }
-  };
-  const validateInteger = (value) => {
-    return Number.isInteger(Number(value));
-  };
-
-  const handleIntegerValidation = (event, key) => {
-    const value = event.target.value;
-    setFormValues((prevState) => ({
-      ...prevState,
-      [key]: value,
-    }));
-
-    if (!validateInteger(value)) {
-      setErrorMessages((prevState) => ({
-        ...prevState,
-        [key]: `${key.replace(/_/g, " ")} must be an integer.`,
-      }));
-    } else {
-      setErrorMessages((prevState) => {
-        const newState = { ...prevState };
-        delete newState[key]; // Delete the error message if the field is valid
-        return newState;
-      });
-    }
-  };
-
-  // If the data is null, show a message indicating that the component has been added
-  if (data === null) {
-    return (
-      <div className="bg-gray-100 p-6">
-        <header className="bg-blue-500 text-white text-center p-4 rounded-md shadow-md">
-          <button
-            onClick={() => removeComponent(id)}
-            className="flex text-red-500"
-          >
-            <FontAwesomeIcon icon={faTrash} />
-          </button>
-          <h1 className="text-3xl font-bold">XREXT Added</h1>
-          <p className="mt-2">The XREXT component has been added successfully.</p>
-        </header>
-      </div>
-    );
-  }
-
-  return (
-    <div className="bg-gray-100 p-6">
-      {/* Header with delete button */}
-      <header className="bg-blue-500 text-white text-center p-4 rounded-md shadow-md">
-        <button
-          onClick={() => removeComponent(id)}
-          className="flex text-red-500"
-        >
-          <FontAwesomeIcon icon={faTrash} />
-        </button>
-        <h1 className="text-3xl font-bold">Xrext Config</h1>
-        <p className="mt-2">Please fill in the fields below to configure the system</p>
-      </header>
-
-      <div className="mt-8 bg-white shadow-md rounded-lg p-6">
-        <form>
-            {/* Additional field 'name' */}
-            <div className="mb-4">
-            <label htmlFor="name" className="block text-gray-700 font-semibold">
-              Name:
-            </label>
-            <input
-              type="text"
-              id={`name-${id}`}
-              name="name"
-              value={formValues.name || ""}
-              onChange={handleChange}
-              className="w-full border border-gray-300 rounded-md p-2 mt-1"
-            />
-            {errorMessages.name && (
-              <small className="block mt-1 text-red-500">{errorMessages.name}</small>
-            )}
-          </div>
-          {Object.keys(data).map((key) => {
-            const field = data[key];
-            if (key === "one_xrext_networks") {
-              return (
-                <div key={key} className="mb-4">
-                  <label className="block text-gray-700 font-semibold mb-2">
-                  {key.replace(/_/g, " ").replace(/\b\w/g, (char) => char.toUpperCase())}:
-                  </label>
-                  {list.map((network, index) => (
-                    <div key={index} className="flex items-center mb-1">
-                      <input
-                        type="checkbox"
-                        id={`${key}_${index}`}
-                        name={key}
-                        value={network}
-                        checked={formValues[key]?.includes(network) || false}
-                        onChange={(event) => handleCheckboxChange(event, key, network)}
-                        className="mr-2"
-                      />
-                      <label htmlFor={`${key}_${index}`} className="text-gray-700">
-                        {network}
-                      </label>
-                    </div>
-                  ))}
-                 <small className="block mt-1 text-gray-500">
-                  {list.length === 0 || list === ""
-                    ? "Create news vnets or tn_vxlan to be able to select"
-                    : "Select one or more networks to include"
-                  }
-                </small>
-                </div>
-              );
-            }
-            return (
-              <div className="mb-4" key={key}>
-                <label htmlFor={key} className="block text-gray-700 font-semibold">
-                  {key.replace(/_/g, " ").replace(/\b\w/g, (char) => char.toUpperCase())}:
-                </label>
-
-                {/* Input or select if there are a 'choices' type */}
-                {field.choices ? (
-                  <select
-                    id={key}
-                    name={key}
-                    value={formValues[key] || ""}
-                    onChange={(event) => handleChange(event)}
-                    className="w-full border border-gray-300 rounded-md p-2 mt-1"
-                  >
-                    <option disabled value="">Select an option</option>
-                    {field.choices.map((choice, index) => (
-                      <option key={index} value={choice}>
-                        {choice}
-                      </option>
-                    ))}
-                  </select>
-                ) : (
-                  <input
-                    type="text"
-                    id={key}
-                    name={key}
-                    value={Array.isArray(formValues[key]) ? formValues[key].join(", ") : formValues[key] || ""}
-                    onChange={(event) => {
-                      if (field.type === "int") {
-                        handleIntegerValidation(event, key);
-                      } else {
-                        handleChange(event);
-                      }
-                    }}
-                    className="w-full border border-gray-300 rounded-md p-2 mt-1"
-                  />
-                )}
-
-                {errorMessages[key] && (
-                  <small className="block mt-1 text-red-500">{errorMessages[key]}</small>
-                )}
-                <small className="block mt-1 text-gray-500">{field.description}</small>
-              </div>
-            );
-          })}
-        </form>
-      </div>
-    </div>
-  );
-};
-
-export default Xrext;
